[GECO-139] Addressing review comments

+++ nepomuk/src/main/java/edu/asu/diging/gilesecosystem/nepomuk/core/service/handlers/AbstractFileHandler.java
@@ -104,13 +104,12 @@ public abstract class AbstractFileHandler implements IFileTypeHandler {
     
     protected abstract IFileStorageManager getStorageManager();
     
-    public void deleteFile(IFile file) {
-        try {
+    public void deleteFile(IFile file) throws NepomukFileStorageException {
+        if (getStorageManager().checkIfFileExists(file.getUsername(), file.getUploadId(), file.getDocumentId(), file.getFilename())) {
             getStorageManager().deleteFile(file.getUsername(), file.getUploadId(), file.getDocumentId(), file.getFilename(), true);
-        } catch (NepomukFileStorageException e) {
-            messageHandler.handleMessage("File could not be deleted.", e, MessageType.WARNING);
         }
-        // files manager will be called to delete the file from the database. The database call will be made even if storage deletion fails since the file could be an older file version that has a database record.
+        // files manager will be called to delete the file from the database. 
+        // Even if the file does not exist in storage as we check above we need to remove the file's database entry as it can be an older file version.
         filesManager.deleteFile(file.getId());
     }
 }
